+++ src/qptr/lower.rs
@@ -8,8 +8,8 @@ use crate::qptr::{shapes, QPtrAttr, QPtrOp};
 use crate::transform::{InnerInPlaceTransform, Transformed, Transformer};
 use crate::{
     spv, AddrSpace, AttrSet, AttrSetDef, Const, ConstCtor, ConstDef, Context, ControlNode,
-    ControlNodeKind, DataInst, DataInstDef, DataInstKind, Diag, FuncDecl, GlobalVarDecl,
-    OrdAssertEq, Type, TypeCtor, TypeCtorArg, TypeDef, Value,
+    ControlNodeKind, DataInst, DataInstDef, DataInstForm, DataInstFormDef, DataInstKind, Diag,
+    FuncDecl, GlobalVarDecl, OrdAssertEq, Type, TypeCtor, TypeCtorArg, TypeDef, Value,
 };
 use smallvec::SmallVec;
 use std::cell::Cell;
@@ -241,6 +241,19 @@ impl Transformer for EraseSpvPtrs<'_> {
             Transformed::Unchanged
         }
     }
+
+    // FIXME(eddyb) because this is now interned, it might be better to
+    // temporarily track the old output types in a map, and not actually
+    // intern the non-`qptr`-output `qptr.*` instructions, only to replace
+    // the output type with `qptr` here.
+    fn transform_data_inst_form_use(
+        &mut self,
+        data_inst_form: DataInstForm,
+    ) -> Transformed<DataInstForm> {
+        // FIXME(eddyb) maybe cache this remap (in `LowerFromSpvPtrs`, globally).
+        self.transform_data_inst_form_def(&self.lowerer.cx[data_inst_form])
+            .map(|data_inst_form_def| self.lowerer.cx.intern(data_inst_form_def))
+    }
 }
 
 struct LowerFromSpvPtrInstsInFunc<'a> {
@@ -414,16 +427,16 @@ impl LowerFromSpvPtrInstsInFunc<'_> {
         // FIXME(eddyb) is this a good convention?
         let func = func_at_data_inst_frozen.at(());
 
-        let spv_inst = match &data_inst_def.kind {
+        let mut attrs = data_inst_def.attrs;
+        let DataInstFormDef {
+            ref kind,
+            output_type,
+        } = cx[data_inst_def.form];
+
+        let spv_inst = match kind {
             DataInstKind::SpvInst(spv_inst) => spv_inst,
             _ => return Ok(Transformed::Unchanged),
         };
-
-        let DataInstDef {
-            mut attrs,
-            output_type,
-            ..
-        } = *data_inst_def;
 
         let replacement_kind_and_inputs = if spv_inst.opcode == wk.OpVariable {
             assert!(data_inst_def.inputs.len() <= 1);
@@ -573,8 +586,10 @@ impl LowerFromSpvPtrInstsInFunc<'_> {
                     cx,
                     DataInstDef {
                         attrs: Default::default(),
-                        kind,
-                        output_type: Some(self.lowerer.qptr_type()),
+                        form: cx.intern(DataInstFormDef {
+                            kind,
+                            output_type: Some(self.lowerer.qptr_type()),
+                        }),
                         inputs,
                     }
                     .into(),
@@ -628,8 +643,13 @@ impl LowerFromSpvPtrInstsInFunc<'_> {
         let (new_kind, new_inputs) = replacement_kind_and_inputs;
         Ok(Transformed::Changed(DataInstDef {
             attrs,
-            kind: new_kind,
-            output_type,
+            // FIXME(eddyb) because this is now interned, it might be better to
+            // temporarily track the old output types in a map, and not actually
+            // intern the non-`qptr`-output `qptr.*` instructions.
+            form: cx.intern(DataInstFormDef {
+                kind: new_kind,
+                output_type,
+            }),
             inputs: new_inputs,
         }))
     }
@@ -643,11 +663,12 @@ impl LowerFromSpvPtrInstsInFunc<'_> {
 
         let func_at_data_inst_frozen = func_at_data_inst.reborrow().freeze();
         let data_inst_def = func_at_data_inst_frozen.def();
+        let data_inst_form_def = &cx[data_inst_def.form];
 
         // FIXME(eddyb) is this a good convention?
         let func = func_at_data_inst_frozen.at(());
 
-        match data_inst_def.kind {
+        match data_inst_form_def.kind {
             // Known semantics, no need to preserve SPIR-V pointer information.
             DataInstKind::FuncCall(_) | DataInstKind::QPtr(_) => return,
 
@@ -673,7 +694,7 @@ impl LowerFromSpvPtrInstsInFunc<'_> {
                     );
             }
         }
-        if let Some(output_type) = data_inst_def.output_type {
+        if let Some(output_type) = data_inst_form_def.output_type {
             if let Some((addr_space, pointee)) = self.lowerer.as_spv_ptr_type(output_type) {
                 old_and_new_attrs
                     .get_or_insert_with(get_old_attrs)